Module type Micheline_sampler.S

Applying the Make functor below yields a module with the following type.

type prim

val sample : (int, prim) Tezos_micheline.Micheline.node Tezos_benchmark.Base_samplers.sampler

sample w is a Micheline sampler for the prescribed primitive type. The sampler uses the provided width function w.