Module Tezos_benchmark_alpha.Michelson_samplers_base

Samplers for basic Michelson values (not including pairs, ors, tickets, big maps, etc)

type parameters = {

int_size : Tezos_benchmark.Base_samplers.range;
(*
The range of the size, measured in bytes, in which big integers must be sampled.
*)

string_size : Tezos_benchmark.Base_samplers.range;
(*
The range of the size, measured in bytes, in which strings must be sampled.
*)

bytes_size : Tezos_benchmark.Base_samplers.range;
(*
The range of the size, measured in bytes, in which bytes must be sampled.
*)

}

Parameters for basic samplers

val parameters_encoding : parameters Data_encoding.t

Encoding for parameters

module type S = sig ... end

A module of type S packs samplers used to construct basic Michelson values.

module Make (P : sig ... end) : S

The Make functor instantiates a module of type S, where the samplers satisfy the given parameters.